set fail_on_uart_buffer_overflow to false on blue2
The issue with the buffer overflows is that the long underwater network
cable is bad, and retransmits happen. The retransmits add around 200ms
in the acquisition loop, which causes the UART on which the sonar
is attached to sometimes overflow. The PTU and pressure sensors are
unaffected as their packets fit in the 256 bytes. The sonar is affected
as its packet is much bigger *and* it sends a regular keepalive.